Fixed some syntax errors

+++ src/racer/player.cr
@@ -1,13 +1,13 @@
 class Player
-  @name:String
-  @avgWPM:Float
-  @wins:Int32
+  @name : String
+  @avgWPM : Float64
+  @wins : Int32
 
   getter name
   getter avgWPM
   getter wins
-  
-  def initialize(properties: Hash(String, Int32 | String | Float64))
+
+  def initialize(properties : Hash(String, Int32 | String | Float64))
       @name = properties["name"]
       @avgWPM = properties["avgWPM"]
       @wins = properties["wins"]
